TST: separate tests of optimize and tnc (moved to test_tnc.py)
